LayoutTests/inspector/canvas/recording-2d-expected.txt

Test that CanvasManager is able to record actions made to 2D canvas contexts.


== Running test suite: Canvas.recording2D
-- Running test case: Canvas.recording2D.singleFrame
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0: (duration)
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc - [native code] (13)
        1: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (15)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (18)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
        5: performActions - inspector/canvas/recording-2d.html:387:25 (22)
        6: Global Code - [program code] (24)
        7: evaluateWithScopeExtension - [native code] (26)
        8: _evaluateOn (28)
        9: _evaluateAndWrap (30)
    1: arc(6, 7, 8, 9, 10, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc - [native code] (13)
        1: (anonymous function) - inspector/canvas/recording-2d.html:78:42 (31)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:78:28 (32)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
        5: performActions - inspector/canvas/recording-2d.html:387:25 (22)
        6: Global Code - [program code] (24)
        7: evaluateWithScopeExtension - [native code] (26)
        8: _evaluateOn (28)
        9: _evaluateAndWrap (30)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "[native code]"
  13: [11,12,0,0]
  14: "inspector/canvas/recording-2d.html"
  15: [10,14,78,42]
  16: "ignoreException"
  17: [16,14,49,13]
  18: [10,14,78,28]
  19: "executeFrameFunction"
  20: [19,14,384,24]
  21: "performActions"
  22: [21,14,388,25]
  23: "global code"
  24: [23,10,1,15]
  25: "evaluateWithScopeExtension"
  26: [25,12,0,0]
  27: "_evaluateOn"
  28: [27,10,130,29]
  29: "_evaluateAndWrap"
  30: [29,10,124,108]
  31: [10,14,79,42]
  32: [10,14,79,28]

-- Running test case: Canvas.recording2D.multipleFrames
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0: (duration)
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc - [native code] (13)
        1: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (15)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (18)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
        5: performActions - inspector/canvas/recording-2d.html:387:25 (22)
        6: Global Code - [program code] (24)
        7: evaluateWithScopeExtension - [native code] (26)
        8: _evaluateOn (28)
        9: _evaluateAndWrap (30)
    1: arc(6, 7, 8, 9, 10, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc - [native code] (13)
        1: (anonymous function) - inspector/canvas/recording-2d.html:78:42 (31)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:78:28 (32)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
        5: performActions - inspector/canvas/recording-2d.html:387:25 (22)
        6: Global Code - [program code] (24)
        7: evaluateWithScopeExtension - [native code] (26)
        8: _evaluateOn (28)
        9: _evaluateAndWrap (30)
  1: (duration)
    0: arcTo(1, 2, 3, 4, 5)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: arcTo - [native code] (34)
        1: (anonymous function) - inspector/canvas/recording-2d.html:81:44 (35)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:81:28 (36)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  2: (duration)
    0: beginPath()
      trace:
        0: beginPath - [native code] (38)
        1: (anonymous function) - inspector/canvas/recording-2d.html:84:26 (39)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  3: (duration)
    0: bezierCurveTo(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: bezierCurveTo - [native code] (41)
        1: (anonymous function) - inspector/canvas/recording-2d.html:87:30 (42)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  4: (duration)
    0: clearRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: clearRect - [native code] (44)
        1: (anonymous function) - inspector/canvas/recording-2d.html:90:26 (45)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  5: (duration)
    0: clearShadow()
      trace:
        0: clearShadow - [native code] (47)
        1: (anonymous function) - inspector/canvas/recording-2d.html:93:28 (48)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  6: (duration)
    0: clip("nonzero")
      swizzleTypes: [String (50)]
      trace:
        0: clip - [native code] (51)
        1: (anonymous function) - inspector/canvas/recording-2d.html:96:21 (52)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: clip("evenodd")
      swizzleTypes: [String (53)]
      trace:
        0: clip - [native code] (51)
        1: (anonymous function) - inspector/canvas/recording-2d.html:97:21 (54)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: clip([object Path2D], "nonzero")
      swizzleTypes: [Path2D (55), String (50)]
      trace:
        0: clip - [native code] (51)
        1: (anonymous function) - inspector/canvas/recording-2d.html:98:21 (56)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: clip([object Path2D], "evenodd")
      swizzleTypes: [Path2D (57), String (53)]
      trace:
        0: clip - [native code] (51)
        1: (anonymous function) - inspector/canvas/recording-2d.html:99:21 (58)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  7: (duration)
    0: closePath()
      trace:
        0: closePath - [native code] (60)
        1: (anonymous function) - inspector/canvas/recording-2d.html:102:26 (61)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  8: (duration)
    0: createImageData([object ImageData])
      swizzleTypes: [ImageData (63)]
      trace:
        0: createImageData - [native code] (64)
        1: (anonymous function) - inspector/canvas/recording-2d.html:105:54 (65)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:105:28 (66)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: createImageData(2, 3)
      swizzleTypes: [Number, Number]
      trace:
        0: createImageData - [native code] (64)
        1: (anonymous function) - inspector/canvas/recording-2d.html:106:54 (67)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:106:28 (68)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  9: (duration)
    0: createLinearGradient(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: createLinearGradient - [native code] (70)
        1: (anonymous function) - inspector/canvas/recording-2d.html:109:59 (71)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:109:28 (72)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  10: (duration)
    0: createPattern([object HTMLImageElement], "testA")
      swizzleTypes: [Image (74), String (75)]
      trace:
        0: createPattern - [native code] (76)
        1: (anonymous function) - inspector/canvas/recording-2d.html:112:52 (77)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:112:28 (78)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: createPattern([object HTMLImageElement], "testB")
      swizzleTypes: [Image (79), String (80)]
      trace:
        0: createPattern - [native code] (76)
        1: (anonymous function) - inspector/canvas/recording-2d.html:113:52 (81)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:113:28 (82)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: createPattern([object HTMLImageElement], "testC")
      swizzleTypes: [Image (83), String (84)]
      trace:
        0: createPattern - [native code] (76)
        1: (anonymous function) - inspector/canvas/recording-2d.html:114:52 (85)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:114:28 (86)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  11: (duration)
    0: createRadialGradient(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: createRadialGradient - [native code] (88)
        1: (anonymous function) - inspector/canvas/recording-2d.html:117:59 (89)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:117:28 (90)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  12: (duration)
    0: direction
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:120:16 (92)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  13: (duration)
    0: drawFocusIfNeeded("Element")
      swizzleTypes: [None (94)]
      trace:
        0: drawFocusIfNeeded - [native code] (95)
        1: (anonymous function) - inspector/canvas/recording-2d.html:124:34 (96)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: drawFocusIfNeeded([object Path2D], "Element")
      swizzleTypes: [Path2D (55), None (94)]
      trace:
        0: drawFocusIfNeeded - [native code] (95)
        1: (anonymous function) - inspector/canvas/recording-2d.html:125:34 (97)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  14: (duration)
    0: drawImage([object HTMLImageElement], 1, 2)
      swizzleTypes: [Image (74), Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:128:48 (100)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:128:28 (101)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: drawImage([object HTMLImageElement], 3, 4)
      swizzleTypes: [Image (79), Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:129:48 (102)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:129:28 (103)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: drawImage([object HTMLImageElement], 5, 6)
      swizzleTypes: [Image (83), Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:130:48 (104)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:130:28 (105)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: drawImage([object HTMLImageElement], 7, 8, 9, 10)
      swizzleTypes: [Image (74), Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:131:48 (106)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:131:28 (107)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: drawImage([object HTMLImageElement], 11, 12, 13, 14)
      swizzleTypes: [Image (79), Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:132:48 (108)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:132:28 (109)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    5: drawImage([object HTMLImageElement], 15, 16, 17, 18)
      swizzleTypes: [Image (83), Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:133:48 (110)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:133:28 (111)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    6: drawImage([object HTMLImageElement], 19, 20, 21, 22, 23, 24, 25, 26)
      swizzleTypes: [Image (74), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:134:48 (112)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:134:28 (113)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    7: drawImage([object HTMLImageElement], 27, 28, 29, 30, 31, 32, 33, 34)
      swizzleTypes: [Image (79), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:135:48 (114)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:135:28 (115)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    8: drawImage([object HTMLImageElement], 35, 36, 37, 38, 39, 40, 41, 42)
      swizzleTypes: [Image (83), Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: drawImage - [native code] (99)
        1: (anonymous function) - inspector/canvas/recording-2d.html:136:48 (116)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:136:28 (117)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  15: (duration)
    0: drawImageFromRect([object HTMLImageElement], 1, 2, 3, 4, 5, 6, 7, 8, "")
      swizzleTypes: [Image (74), Number, Number, Number, Number, Number, Number, Number, Number, String (10)]
      trace:
        0: drawImageFromRect - [native code] (119)
        1: (anonymous function) - inspector/canvas/recording-2d.html:139:34 (120)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: drawImageFromRect([object HTMLImageElement], 9, 10, 11, 12, 13, 14, 15, 16, "test")
      swizzleTypes: [Image (74), Number, Number, Number, Number, Number, Number, Number, Number, String (121)]
      trace:
        0: drawImageFromRect - [native code] (119)
        1: (anonymous function) - inspector/canvas/recording-2d.html:140:34 (122)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  16: (duration)
    0: ellipse(1, 2, 3, 4, 5, 6, 7, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Boolean]
      trace:
        0: ellipse - [native code] (124)
        1: (anonymous function) - inspector/canvas/recording-2d.html:143:46 (125)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:143:28 (126)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: ellipse(8, 9, 10, 11, 12, 13, 14, true)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Boolean]
      trace:
        0: ellipse - [native code] (124)
        1: (anonymous function) - inspector/canvas/recording-2d.html:144:46 (127)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:144:28 (128)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  17: (duration)
    0: fill("nonzero")
      swizzleTypes: [String (50)]
      trace:
        0: fill - [native code] (130)
        1: (anonymous function) - inspector/canvas/recording-2d.html:147:21 (131)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: fill("evenodd")
      swizzleTypes: [String (53)]
      trace:
        0: fill - [native code] (130)
        1: (anonymous function) - inspector/canvas/recording-2d.html:148:21 (132)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: fill([object Path2D], "nonzero")
      swizzleTypes: [Path2D (55), String (50)]
      trace:
        0: fill - [native code] (130)
        1: (anonymous function) - inspector/canvas/recording-2d.html:149:21 (133)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: fill([object Path2D], "evenodd")
      swizzleTypes: [Path2D (57), String (53)]
      trace:
        0: fill - [native code] (130)
        1: (anonymous function) - inspector/canvas/recording-2d.html:150:21 (134)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  18: (duration)
    0: fillRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: fillRect - [native code] (136)
        1: (anonymous function) - inspector/canvas/recording-2d.html:153:25 (137)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  19: (duration)
    0: fillStyle
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:156:16 (139)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: fillStyle = "test"
      swizzleTypes: [String (121)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:157:16 (140)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: fillStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (142)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:158:16 (143)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: fillStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (145)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:159:16 (146)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: fillStyle = [object CanvasPattern]
      swizzleTypes: [CanvasPattern (149)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:160:16 (150)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  20: (duration)
    0: fillText("testA", 1, 2)
      swizzleTypes: [String (75), Number, Number]
      trace:
        0: fillText - [native code] (152)
        1: (anonymous function) - inspector/canvas/recording-2d.html:163:25 (153)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: fillText("testB", 3, 4, 5)
      swizzleTypes: [String (80), Number, Number, Number]
      trace:
        0: fillText - [native code] (152)
        1: (anonymous function) - inspector/canvas/recording-2d.html:164:25 (154)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  21: (duration)
    0: font
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:167:16 (156)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: font = "test"
      swizzleTypes: [String (121)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:168:16 (157)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  22: (duration)
    0: getImageData(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: getImageData - [native code] (159)
        1: (anonymous function) - inspector/canvas/recording-2d.html:171:51 (160)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:171:28 (161)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  23: (duration)
    0: getLineDash()
      trace:
        0: getLineDash - [native code] (163)
        1: (anonymous function) - inspector/canvas/recording-2d.html:174:28 (164)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  24: (duration)
    0: getTransform
      trace:
        0: getTransform - [native code] (166)
        1: (anonymous function) - inspector/canvas/recording-2d.html:177:29 (167)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  25: (duration)
    0: globalAlpha
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:180:16 (169)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: globalAlpha = 0
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:181:16 (170)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  26: (duration)
    0: globalCompositeOperation
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:184:16 (172)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: globalCompositeOperation = "test"
      swizzleTypes: [String (121)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:185:16 (173)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  27: (duration)
    0: imageSmoothingEnabled
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:188:16 (175)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: imageSmoothingEnabled = true
      swizzleTypes: [Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:189:16 (176)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  28: (duration)
    0: imageSmoothingQuality
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:192:16 (178)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: imageSmoothingQuality = "low"
      swizzleTypes: [String (9)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:193:16 (179)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  29: (duration)
    0: isPointInPath([object Path2D], 5, 6, "nonzero")
      swizzleTypes: [Path2D (55), Number, Number, String (50)]
      trace:
        0: isPointInPath - [native code] (181)
        1: (anonymous function) - inspector/canvas/recording-2d.html:196:30 (182)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: isPointInPath([object Path2D], 7, 8, "evenodd")
      swizzleTypes: [Path2D (57), Number, Number, String (53)]
      trace:
        0: isPointInPath - [native code] (181)
        1: (anonymous function) - inspector/canvas/recording-2d.html:197:30 (183)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: isPointInPath(9, 10, "nonzero")
      swizzleTypes: [Number, Number, String (50)]
      trace:
        0: isPointInPath - [native code] (181)
        1: (anonymous function) - inspector/canvas/recording-2d.html:198:30 (184)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: isPointInPath(11, 12, "evenodd")
      swizzleTypes: [Number, Number, String (53)]
      trace:
        0: isPointInPath - [native code] (181)
        1: (anonymous function) - inspector/canvas/recording-2d.html:199:30 (185)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  30: (duration)
    0: isPointInStroke([object Path2D], 3, 4)
      swizzleTypes: [Path2D (55), Number, Number]
      trace:
        0: isPointInStroke - [native code] (187)
        1: (anonymous function) - inspector/canvas/recording-2d.html:202:32 (188)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: isPointInStroke(5, 6)
      swizzleTypes: [Number, Number]
      trace:
        0: isPointInStroke - [native code] (187)
        1: (anonymous function) - inspector/canvas/recording-2d.html:203:32 (189)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  31: (duration)
    0: lineCap
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:206:16 (191)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  32: (duration)
    0: lineDashOffset
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:210:16 (193)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: lineDashOffset = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:211:16 (194)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  33: (duration)
    0: lineJoin
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:214:16 (196)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  34: (duration)
    0: lineTo(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: lineTo - [native code] (198)
        1: (anonymous function) - inspector/canvas/recording-2d.html:218:23 (199)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  35: (duration)
    0: lineWidth
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:221:16 (201)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: lineWidth = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:222:16 (202)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  36: (duration)
    0: measureText("test")
      swizzleTypes: [String (121)]
      trace:
        0: measureText - [native code] (204)
        1: (anonymous function) - inspector/canvas/recording-2d.html:225:28 (205)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  37: (duration)
    0: miterLimit
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:228:16 (207)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: miterLimit = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:229:16 (208)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  38: (duration)
    0: moveTo(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: moveTo - [native code] (210)
        1: (anonymous function) - inspector/canvas/recording-2d.html:232:23 (211)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  39: (duration)
    0: putImageData([object ImageData], 5, 6)
      swizzleTypes: [ImageData (63), Number, Number]
      trace:
        0: putImageData - [native code] (213)
        1: (anonymous function) - inspector/canvas/recording-2d.html:235:29 (214)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: putImageData([object ImageData], 7, 8, 9, 10, 11, 12)
      swizzleTypes: [ImageData (215), Number, Number, Number, Number, Number, Number]
      trace:
        0: putImageData - [native code] (213)
        1: (anonymous function) - inspector/canvas/recording-2d.html:236:29 (216)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  40: (duration)
    0: quadraticCurveTo(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: quadraticCurveTo - [native code] (218)
        1: (anonymous function) - inspector/canvas/recording-2d.html:239:33 (219)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  41: (duration)
    0: rect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: rect - [native code] (221)
        1: (anonymous function) - inspector/canvas/recording-2d.html:242:21 (222)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  42: (duration)
    0: resetTransform()
      trace:
        0: resetTransform - [native code] (224)
        1: (anonymous function) - inspector/canvas/recording-2d.html:245:31 (225)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  43: (duration)
    0: restore()
      trace:
        0: restore - [native code] (227)
        1: (anonymous function) - inspector/canvas/recording-2d.html:248:24 (228)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  44: (duration)
    0: rotate(1)
      swizzleTypes: [Number]
      trace:
        0: rotate - [native code] (230)
        1: (anonymous function) - inspector/canvas/recording-2d.html:253:23 (231)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  45: (duration)
    0: save()
      trace:
        0: save - [native code] (233)
        1: (anonymous function) - inspector/canvas/recording-2d.html:256:21 (234)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  46: (duration)
    0: scale(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: scale - [native code] (236)
        1: (anonymous function) - inspector/canvas/recording-2d.html:259:22 (237)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  47: (duration)
    0: setAlpha(null)
      swizzleTypes: [Number]
      trace:
        0: setAlpha - [native code] (239)
        1: (anonymous function) - inspector/canvas/recording-2d.html:262:25 (240)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setAlpha(1)
      swizzleTypes: [Number]
      trace:
        0: setAlpha - [native code] (239)
        1: (anonymous function) - inspector/canvas/recording-2d.html:263:25 (241)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  48: (duration)
    0: setCompositeOperation("undefined")
      swizzleTypes: [String (243)]
      trace:
        0: setCompositeOperation - [native code] (244)
        1: (anonymous function) - inspector/canvas/recording-2d.html:266:38 (245)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setCompositeOperation("test")
      swizzleTypes: [String (121)]
      trace:
        0: setCompositeOperation - [native code] (244)
        1: (anonymous function) - inspector/canvas/recording-2d.html:267:38 (246)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  49: (duration)
    0: setFillColor("testA")
      swizzleTypes: [String (75)]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:270:29 (249)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setFillColor("testB", 1)
      swizzleTypes: [String (80), Number]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:271:29 (250)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: setFillColor(2, 1)
      swizzleTypes: [Number, Number]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:272:29 (251)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: setFillColor(3, 4)
      swizzleTypes: [Number, Number]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:273:29 (252)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: setFillColor(5, 6, 7, 8)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:274:29 (253)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    5: setFillColor(9, 10, 11, 12, 13)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: setFillColor - [native code] (248)
        1: (anonymous function) - inspector/canvas/recording-2d.html:275:29 (254)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  50: (duration)
    0: setLineCap("undefined")
      swizzleTypes: [String (243)]
      trace:
        0: setLineCap - [native code] (256)
        1: (anonymous function) - inspector/canvas/recording-2d.html:278:27 (257)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setLineCap("test")
      swizzleTypes: [String (121)]
      trace:
        0: setLineCap - [native code] (256)
        1: (anonymous function) - inspector/canvas/recording-2d.html:279:27 (258)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  51: (duration)
    0: setLineDash([1,2])
      swizzleTypes: [Array]
      trace:
        0: setLineDash - [native code] (260)
        1: (anonymous function) - inspector/canvas/recording-2d.html:282:28 (261)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  52: (duration)
    0: setLineJoin("undefined")
      swizzleTypes: [String (243)]
      trace:
        0: setLineJoin - [native code] (263)
        1: (anonymous function) - inspector/canvas/recording-2d.html:285:28 (264)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setLineJoin("test")
      swizzleTypes: [String (121)]
      trace:
        0: setLineJoin - [native code] (263)
        1: (anonymous function) - inspector/canvas/recording-2d.html:286:28 (265)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  53: (duration)
    0: setLineWidth(null)
      swizzleTypes: [Number]
      trace:
        0: setLineWidth - [native code] (267)
        1: (anonymous function) - inspector/canvas/recording-2d.html:289:29 (268)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setLineWidth(1)
      swizzleTypes: [Number]
      trace:
        0: setLineWidth - [native code] (267)
        1: (anonymous function) - inspector/canvas/recording-2d.html:290:29 (269)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  54: (duration)
    0: setMiterLimit(null)
      swizzleTypes: [Number]
      trace:
        0: setMiterLimit - [native code] (271)
        1: (anonymous function) - inspector/canvas/recording-2d.html:293:30 (272)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setMiterLimit(1)
      swizzleTypes: [Number]
      trace:
        0: setMiterLimit - [native code] (271)
        1: (anonymous function) - inspector/canvas/recording-2d.html:294:30 (273)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  55: (duration)
    0: setShadow(1, 2, 3, "")
      swizzleTypes: [Number, Number, Number, String (275)]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:297:26 (277)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setShadow(4, 5, 6, "test", 7)
      swizzleTypes: [Number, Number, Number, String (121), Number]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:298:26 (278)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: setShadow(8, 9, 10, 11, 1)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:299:26 (279)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: setShadow(12, 13, 14, 15, 16)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:300:26 (280)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: setShadow(17, 18, 19, 20, 21, 22, 23)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:301:26 (281)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    5: setShadow(24, 25, 26, 27, 28, 29, 30, 31)
      swizzleTypes: [Number, Number, Number, Number, Number, Number, Number, Number]
      trace:
        0: setShadow - [native code] (276)
        1: (anonymous function) - inspector/canvas/recording-2d.html:302:26 (282)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  56: (duration)
    0: setStrokeColor("testA")
      swizzleTypes: [String (75)]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:305:31 (285)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setStrokeColor("testB", 1)
      swizzleTypes: [String (80), Number]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:306:31 (286)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: setStrokeColor(2, 1)
      swizzleTypes: [Number, Number]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:307:31 (287)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: setStrokeColor(3, 4)
      swizzleTypes: [Number, Number]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:308:31 (288)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: setStrokeColor(5, 6, 7, 8)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:309:31 (289)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    5: setStrokeColor(9, 10, 11, 12, 13)
      swizzleTypes: [Number, Number, Number, Number, Number]
      trace:
        0: setStrokeColor - [native code] (284)
        1: (anonymous function) - inspector/canvas/recording-2d.html:310:31 (290)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  57: (duration)
    0: setTransform(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: setTransform - [native code] (292)
        1: (anonymous function) - inspector/canvas/recording-2d.html:313:29 (293)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: setTransform(matrix(1, 0, 0, 1, 0, 0))
      swizzleTypes: [DOMMatrix]
      trace:
        0: setTransform - [native code] (292)
        1: (anonymous function) - inspector/canvas/recording-2d.html:314:51 (294)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:314:28 (295)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: setTransform(matrix(7, 8, 9, 10, 11, 12))
      swizzleTypes: [DOMMatrix]
      trace:
        0: setTransform - [native code] (292)
        1: (anonymous function) - inspector/canvas/recording-2d.html:315:51 (296)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:315:28 (297)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  58: (duration)
    0: shadowBlur
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:318:16 (299)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: shadowBlur = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:319:16 (300)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  59: (duration)
    0: shadowColor
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:322:16 (302)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: shadowColor = "test"
      swizzleTypes: [String (121)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:323:16 (303)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  60: (duration)
    0: shadowOffsetX
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:326:16 (305)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: shadowOffsetX = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:327:16 (306)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  61: (duration)
    0: shadowOffsetY
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:330:16 (308)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: shadowOffsetY = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:331:16 (309)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  62: (duration)
    0: stroke()
      trace:
        0: stroke - [native code] (311)
        1: (anonymous function) - inspector/canvas/recording-2d.html:334:23 (312)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: stroke([object Path2D])
      swizzleTypes: [Path2D (55)]
      trace:
        0: stroke - [native code] (311)
        1: (anonymous function) - inspector/canvas/recording-2d.html:335:23 (313)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  63: (duration)
    0: strokeRect(1, 2, 3, 4)
      swizzleTypes: [Number, Number, Number, Number]
      trace:
        0: strokeRect - [native code] (315)
        1: (anonymous function) - inspector/canvas/recording-2d.html:338:27 (316)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  64: (duration)
    0: strokeStyle
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:341:16 (318)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: strokeStyle = "test"
      swizzleTypes: [String (121)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:342:16 (319)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    2: strokeStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (142)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:343:16 (320)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    3: strokeStyle = [object CanvasGradient]
      swizzleTypes: [CanvasGradient (145)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:344:16 (321)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    4: strokeStyle = [object CanvasPattern]
      swizzleTypes: [CanvasPattern (149)]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:345:16 (322)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  65: (duration)
    0: strokeText("testA", 1, 2)
      swizzleTypes: [String (75), Number, Number]
      trace:
        0: strokeText - [native code] (324)
        1: (anonymous function) - inspector/canvas/recording-2d.html:348:27 (325)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: strokeText("testB", 3, 4, 5)
      swizzleTypes: [String (80), Number, Number, Number]
      trace:
        0: strokeText - [native code] (324)
        1: (anonymous function) - inspector/canvas/recording-2d.html:349:27 (326)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  66: (duration)
    0: textAlign
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:352:16 (328)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  67: (duration)
    0: textBaseline
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:356:16 (330)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  68: (duration)
    0: transform(1, 2, 3, 4, 5, 6)
      swizzleTypes: [Number, Number, Number, Number, Number, Number]
      trace:
        0: transform - [native code] (332)
        1: (anonymous function) - inspector/canvas/recording-2d.html:360:26 (333)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  69: (duration)
    0: translate(1, 2)
      swizzleTypes: [Number, Number]
      trace:
        0: translate - [native code] (335)
        1: (anonymous function) - inspector/canvas/recording-2d.html:363:26 (336)
        2: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  70: (duration)
    0: webkitImageSmoothingEnabled
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:366:16 (338)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: webkitImageSmoothingEnabled = true
      swizzleTypes: [Boolean]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:367:16 (339)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  71: (duration)
    0: webkitLineDash
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:370:16 (341)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: webkitLineDash = [1,2]
      swizzleTypes: [Array]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:371:16 (342)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
  72: (duration)
    0: webkitLineDashOffset
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:374:16 (344)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
    1: webkitLineDashOffset = 1
      swizzleTypes: [Number]
      trace:
        0: (anonymous function) - inspector/canvas/recording-2d.html:375:16 (345)
        1: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "[native code]"
  13: [11,12,0,0]
  14: "inspector/canvas/recording-2d.html"
  15: [10,14,78,42]
  16: "ignoreException"
  17: [16,14,49,13]
  18: [10,14,78,28]
  19: "executeFrameFunction"
  20: [19,14,384,24]
  21: "performActions"
  22: [21,14,388,25]
  23: "global code"
  24: [23,10,1,15]
  25: "evaluateWithScopeExtension"
  26: [25,12,0,0]
  27: "_evaluateOn"
  28: [27,10,130,29]
  29: "_evaluateAndWrap"
  30: [29,10,124,108]
  31: [10,14,79,42]
  32: [10,14,79,28]
  33: "arcTo"
  34: [33,12,0,0]
  35: [10,14,82,44]
  36: [10,14,82,28]
  37: "beginPath"
  38: [37,12,0,0]
  39: [10,14,85,26]
  40: "bezierCurveTo"
  41: [40,12,0,0]
  42: [10,14,88,30]
  43: "clearRect"
  44: [43,12,0,0]
  45: [10,14,91,26]
  46: "clearShadow"
  47: [46,12,0,0]
  48: [10,14,94,28]
  49: "clip"
  50: "nonzero"
  51: [49,12,0,0]
  52: [10,14,97,21]
  53: "evenodd"
  54: [10,14,98,21]
  55: "M1 2"
  56: [10,14,99,21]
  57: "M3 4"
  58: [10,14,100,21]
  59: "closePath"
  60: [59,12,0,0]
  61: [10,14,103,26]
  62: "createImageData"
  63: [[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],1,4]
  64: [62,12,0,0]
  65: [10,14,106,54]
  66: [10,14,106,28]
  67: [10,14,107,54]
  68: [10,14,107,28]
  69: "createLinearGradient"
  70: [69,12,0,0]
  71: [10,14,110,59]
  72: [10,14,110,28]
  73: "createPattern"
  74: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAABNJREFUCB1j/M/AAEQMDEwgAgQAHxcCAmtAm/sAAAAASUVORK5CYII="
  75: "testA"
  76: [73,12,0,0]
  77: [10,14,113,52]
  78: [10,14,113,28]
  79: "data:,"
  80: "testB"
  81: [10,14,114,52]
  82: [10,14,114,28]
  83: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
  84: "testC"
  85: [10,14,115,52]
  86: [10,14,115,28]
  87: "createRadialGradient"
  88: [87,12,0,0]
  89: [10,14,118,59]
  90: [10,14,118,28]
  91: "direction"
  92: [10,14,121,16]
  93: "drawFocusIfNeeded"
  94: "Element"
  95: [93,12,0,0]
  96: [10,14,125,34]
  97: [10,14,126,34]
  98: "drawImage"
  99: [98,12,0,0]
  100: [10,14,129,48]
  101: [10,14,129,28]
  102: [10,14,130,48]
  103: [10,14,130,28]
  104: [10,14,131,48]
  105: [10,14,131,28]
  106: [10,14,132,48]
  107: [10,14,132,28]
  108: [10,14,133,48]
  109: [10,14,133,28]
  110: [10,14,134,48]
  111: [10,14,134,28]
  112: [10,14,135,48]
  113: [10,14,135,28]
  114: [10,14,136,48]
  115: [10,14,136,28]
  116: [10,14,137,48]
  117: [10,14,137,28]
  118: "drawImageFromRect"
  119: [118,12,0,0]
  120: [10,14,140,34]
  121: "test"
  122: [10,14,141,34]
  123: "ellipse"
  124: [123,12,0,0]
  125: [10,14,144,46]
  126: [10,14,144,28]
  127: [10,14,145,46]
  128: [10,14,145,28]
  129: "fill"
  130: [129,12,0,0]
  131: [10,14,148,21]
  132: [10,14,149,21]
  133: [10,14,150,21]
  134: [10,14,151,21]
  135: "fillRect"
  136: [135,12,0,0]
  137: [10,14,154,25]
  138: "fillStyle"
  139: [10,14,157,16]
  140: [10,14,158,16]
  141: "linear-gradient"
  142: [141,[1,2,3,4],[]]
  143: [10,14,159,16]
  144: "radial-gradient"
  145: [144,[1,2,3,4,5,6],[]]
  146: [10,14,160,16]
  147: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAABNJREFUCB1j/M/AAEQMDEwgAgQAHxcCAmtAm/sAAAAASUVORK5CYII="
  148: "no-repeat"
  149: [147,148]
  150: [10,14,161,16]
  151: "fillText"
  152: [151,12,0,0]
  153: [10,14,164,25]
  154: [10,14,165,25]
  155: "font"
  156: [10,14,168,16]
  157: [10,14,169,16]
  158: "getImageData"
  159: [158,12,0,0]
  160: [10,14,172,51]
  161: [10,14,172,28]
  162: "getLineDash"
  163: [162,12,0,0]
  164: [10,14,175,28]
  165: "getTransform"
  166: [165,12,0,0]
  167: [10,14,178,29]
  168: "globalAlpha"
  169: [10,14,181,16]
  170: [10,14,182,16]
  171: "globalCompositeOperation"
  172: [10,14,185,16]
  173: [10,14,186,16]
  174: "imageSmoothingEnabled"
  175: [10,14,189,16]
  176: [10,14,190,16]
  177: "imageSmoothingQuality"
  178: [10,14,193,16]
  179: [10,14,194,16]
  180: "isPointInPath"
  181: [180,12,0,0]
  182: [10,14,197,30]
  183: [10,14,198,30]
  184: [10,14,199,30]
  185: [10,14,200,30]
  186: "isPointInStroke"
  187: [186,12,0,0]
  188: [10,14,203,32]
  189: [10,14,204,32]
  190: "lineCap"
  191: [10,14,207,16]
  192: "lineDashOffset"
  193: [10,14,211,16]
  194: [10,14,212,16]
  195: "lineJoin"
  196: [10,14,215,16]
  197: "lineTo"
  198: [197,12,0,0]
  199: [10,14,219,23]
  200: "lineWidth"
  201: [10,14,222,16]
  202: [10,14,223,16]
  203: "measureText"
  204: [203,12,0,0]
  205: [10,14,226,28]
  206: "miterLimit"
  207: [10,14,229,16]
  208: [10,14,230,16]
  209: "moveTo"
  210: [209,12,0,0]
  211: [10,14,233,23]
  212: "putImageData"
  213: [212,12,0,0]
  214: [10,14,236,29]
  215: [[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],2,3]
  216: [10,14,237,29]
  217: "quadraticCurveTo"
  218: [217,12,0,0]
  219: [10,14,240,33]
  220: "rect"
  221: [220,12,0,0]
  222: [10,14,243,21]
  223: "resetTransform"
  224: [223,12,0,0]
  225: [10,14,246,31]
  226: "restore"
  227: [226,12,0,0]
  228: [10,14,249,24]
  229: "rotate"
  230: [229,12,0,0]
  231: [10,14,254,23]
  232: "save"
  233: [232,12,0,0]
  234: [10,14,257,21]
  235: "scale"
  236: [235,12,0,0]
  237: [10,14,260,22]
  238: "setAlpha"
  239: [238,12,0,0]
  240: [10,14,263,25]
  241: [10,14,264,25]
  242: "setCompositeOperation"
  243: "undefined"
  244: [242,12,0,0]
  245: [10,14,267,38]
  246: [10,14,268,38]
  247: "setFillColor"
  248: [247,12,0,0]
  249: [10,14,271,29]
  250: [10,14,272,29]
  251: [10,14,273,29]
  252: [10,14,274,29]
  253: [10,14,275,29]
  254: [10,14,276,29]
  255: "setLineCap"
  256: [255,12,0,0]
  257: [10,14,279,27]
  258: [10,14,280,27]
  259: "setLineDash"
  260: [259,12,0,0]
  261: [10,14,283,28]
  262: "setLineJoin"
  263: [262,12,0,0]
  264: [10,14,286,28]
  265: [10,14,287,28]
  266: "setLineWidth"
  267: [266,12,0,0]
  268: [10,14,290,29]
  269: [10,14,291,29]
  270: "setMiterLimit"
  271: [270,12,0,0]
  272: [10,14,294,30]
  273: [10,14,295,30]
  274: "setShadow"
  275: ""
  276: [274,12,0,0]
  277: [10,14,298,26]
  278: [10,14,299,26]
  279: [10,14,300,26]
  280: [10,14,301,26]
  281: [10,14,302,26]
  282: [10,14,303,26]
  283: "setStrokeColor"
  284: [283,12,0,0]
  285: [10,14,306,31]
  286: [10,14,307,31]
  287: [10,14,308,31]
  288: [10,14,309,31]
  289: [10,14,310,31]
  290: [10,14,311,31]
  291: "setTransform"
  292: [291,12,0,0]
  293: [10,14,314,29]
  294: [10,14,315,51]
  295: [10,14,315,28]
  296: [10,14,316,51]
  297: [10,14,316,28]
  298: "shadowBlur"
  299: [10,14,319,16]
  300: [10,14,320,16]
  301: "shadowColor"
  302: [10,14,323,16]
  303: [10,14,324,16]
  304: "shadowOffsetX"
  305: [10,14,327,16]
  306: [10,14,328,16]
  307: "shadowOffsetY"
  308: [10,14,331,16]
  309: [10,14,332,16]
  310: "stroke"
  311: [310,12,0,0]
  312: [10,14,335,23]
  313: [10,14,336,23]
  314: "strokeRect"
  315: [314,12,0,0]
  316: [10,14,339,27]
  317: "strokeStyle"
  318: [10,14,342,16]
  319: [10,14,343,16]
  320: [10,14,344,16]
  321: [10,14,345,16]
  322: [10,14,346,16]
  323: "strokeText"
  324: [323,12,0,0]
  325: [10,14,349,27]
  326: [10,14,350,27]
  327: "textAlign"
  328: [10,14,353,16]
  329: "textBaseline"
  330: [10,14,357,16]
  331: "transform"
  332: [331,12,0,0]
  333: [10,14,361,26]
  334: "translate"
  335: [334,12,0,0]
  336: [10,14,364,26]
  337: "webkitImageSmoothingEnabled"
  338: [10,14,367,16]
  339: [10,14,368,16]
  340: "webkitLineDash"
  341: [10,14,371,16]
  342: [10,14,372,16]
  343: "webkitLineDashOffset"
  344: [10,14,375,16]
  345: [10,14,376,16]

-- Running test case: Canvas.recording2D.memoryLimit
initialState:
  attributes:
    width: 2
    height: 2
    setTransform: [1,0,0,1,0,0]
    globalAlpha: 1
    globalCompositeOperation: 0
    lineWidth: 1
    lineCap: 1
    lineJoin: 2
    miterLimit: 10
    shadowOffsetX: 0
    shadowOffsetY: 0
    shadowBlur: 0
    shadowColor: 3
    setLineDash: [[]]
    lineDashOffset: 0
    font: 4
    textAlign: 5
    textBaseline: 6
    direction: 7
    strokeStyle: 8
    fillStyle: 8
    imageSmoothingEnabled: true
    imageSmoothingQuality: 9
    setPath: [10]
  parameters:
  content: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAYAAABytg0kAAAAAXNSR0IArs4c6QAAAAtJREFUCB1jYEAHAAASAAGAFMrMAAAAAElFTkSuQmCC"
frames:
  0: (duration) (incomplete)
    0: arc(1, 2, 3, 4, 5, false)
      swizzleTypes: [Number, Number, Number, Number, Number, Boolean]
      trace:
        0: arc - [native code] (13)
        1: (anonymous function) - inspector/canvas/recording-2d.html:77:42 (15)
        2: ignoreException - inspector/canvas/recording-2d.html:48:13 (17)
        3: (anonymous function) - inspector/canvas/recording-2d.html:77:28 (18)
        4: executeFrameFunction - inspector/canvas/recording-2d.html:383:24 (20)
        5: performActions - inspector/canvas/recording-2d.html:387:25 (22)
        6: Global Code - [program code] (24)
        7: evaluateWithScopeExtension - [native code] (26)
        8: _evaluateOn (28)
        9: _evaluateAndWrap (30)
data:
  0: "source-over"
  1: "butt"
  2: "miter"
  3: "rgba(0, 0, 0, 0)"
  4: "10px sans-serif"
  5: "start"
  6: "alphabetic"
  7: "ltr"
  8: "#000000"
  9: "low"
  10: ""
  11: "arc"
  12: "[native code]"
  13: [11,12,0,0]
  14: "inspector/canvas/recording-2d.html"
  15: [10,14,78,42]
  16: "ignoreException"
  17: [16,14,49,13]
  18: [10,14,78,28]
  19: "executeFrameFunction"
  20: [19,14,384,24]
  21: "performActions"
  22: [21,14,388,25]
  23: "global code"
  24: [23,10,1,15]
  25: "evaluateWithScopeExtension"
  26: [25,12,0,0]
  27: "_evaluateOn"
  28: [27,10,130,29]
  29: "_evaluateAndWrap"
  30: [29,10,124,108]

-- Running test case: Canvas.recording2D.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.

-- Running test case: Canvas.recording2D.ActionParameterNaN
PASS: The recording should have 1 frame.
PASS: The first frame should have 1 action.
PASS: The action should have 1 parameter.
PASS: The parameter should be null.

-- Running test case: Canvas.recording2D.NoActions
PASS: A recording should have been started and stopped once.
PASS: A recording should have been started and stopped twice.
